third_party/blink/web_tests/svg/dynamic-updates/SVGFilterElement-svgdom-filterUnits-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'filterUnits' property of the SVGFilterElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS filterElement.filterUnits.baseVal is SVGUnitTypes.SVG_UNIT_TYPE_USERSPACEONUSE
PASS filterElement.filterUnits.baseVal is SVGUnitTypes.SVG_UNIT_TYPE_OBJECTBOUNDINGBOX
PASS successfullyParsed is true

TEST COMPLETE
